backend/requirements.txt
celery~=5.1.2
django~=3.2.8
djangorestframework~=3.12.4
python-dotenv~=0.19.1
requests~=2.26.0
web3~=5.24.0
